+++ src/main/java/by/project/trucking_v2/service/OrderServiceImpl.java
@@ -1,11 +1,10 @@
 package by.project.trucking_v2.service;
 
+import by.project.trucking_v2.exception.AccessDeniedException;
+import by.project.trucking_v2.exception.EmptyResultException;
 import by.project.trucking_v2.exception.NotFoundException;
 import by.project.trucking_v2.model.*;
-import by.project.trucking_v2.repository.CompletedOrderRepository;
-import by.project.trucking_v2.repository.LegalEntityRepository;
-import by.project.trucking_v2.repository.OrderRepository;
-import by.project.trucking_v2.repository.TransportRepository;
+import by.project.trucking_v2.repository.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,21 +15,17 @@ import java.util.List;
 
 @Service
 public class OrderServiceImpl implements OrderService {
-    public static final String ANSI_RED = "\u001B[31m";
-    public static final String ANSI_YELLOW = "\u001B[33m";
-    public static final String ANSI_GREEN = "\u001B[32m";
-    public static final String ANSI_RESET = "\u001B[0m";
+
     private static final Logger log = LoggerFactory.getLogger(OrderServiceImpl.class);
 
     @Autowired
-    OrderRepository orderRepository;
+    private OrderRepository orderRepository;
     @Autowired
-    LegalEntityRepository legalEntityRepository;
-
+    private LegalEntityRepository legalEntityRepository;
     @Autowired
-    CompletedOrderRepository completedOrderRepository;
+    private CompletedOrderServiceImpl completedOrderService;
     @Autowired
-    CompletedOrderServiceImpl completedOrderService;
+    private UserRepository userRepository;
 
     @Override
     public List<Order> getAll() {
@@ -40,27 +35,33 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Order findById(int id) {
+    public Order findById(Integer id) {
         return orderRepository.findById(id).orElseThrow(() -> new NotFoundException());
     }
 
     @Override
-    public Order save(Order order, int le_id) {
-        LegalEntity le = legalEntityRepository.findById(le_id).orElseThrow(NotFoundException::new);
-        Order newOrder = new Order();
-        newOrder.setDescription(order.getDescription());
-        newOrder.setWeight(order.getWeight());
-        newOrder.setDeparture(order.getDeparture());
-        newOrder.setDestination(order.getDestination());
-        newOrder.setPrice(order.getPrice());
-        newOrder.setStatus(order.getStatus());
-        newOrder.setLegalEntity(le);
-        return orderRepository.save(newOrder);
+    public Order save(Order order) {
+        /** проверка на наличие юр.лица */
+        if (legalEntityRepository.existsById(order.getLegalEntity().getId())) {
+
+            /** проверка роли */
+            User u = userRepository.findByLegalEntityId(order.getLegalEntity().getId()); //нах юзера по легал энтити айди
+            if (u.getRole() == Role.CLIENT) {
+                log.info("Заказ успешно сохранен");
+                return orderRepository.save(order);
+            } else {
+                log.warn("Заказ НЕ сохранен. Вы должны быть зарегистрированы как клиент");
+                throw new AccessDeniedException();
+            }
+        } else {
+            log.warn("Заказ НЕ сохранен. Юр.лицо с id='" + order.getLegalEntity().getId() + "' не найдено");
+            throw new NotFoundException();
+        }
     }
 
 
     @Override
-    public Order update(int id, Order order) {
+    public Order update(Integer id, Order order) {
         Order currentOrder = orderRepository.findById(id).orElseThrow(NotFoundException::new);
         currentOrder.setDescription(order.getDescription());
         currentOrder.setWeight(order.getWeight());
@@ -73,19 +74,29 @@ public class OrderServiceImpl implements OrderService {
        }
 
     @Override
-    public void delete(int id) {
-        orderRepository.deleteById(id);
+    public void delete(Integer id) {
+        if (orderRepository.existsById(id)) {
+            orderRepository.deleteById(id);
+            log.info("Заказ успешно удален");
+        } else {
+            log.warn("Заказ НЕ удален. Заказа с id=" + id + " не существует");
+            throw new EmptyResultException();
+        }
     }
 
-
     @Override
-    public Order orderChoice(int id, Order order) {
+    public Order orderChoice(Integer id, Order order) {
         Order currentOrder = orderRepository.findById(id).orElseThrow(NotFoundException::new);
-        currentOrder.setStatus(Status.IN_PROGRESS);
+        if (currentOrder.getStatus() == Status.AVALIABLE) {
 
-        completedOrderService.orderInProgress(id);
+            currentOrder.setStatus(Status.IN_PROGRESS);
+            completedOrderService.orderInProgress(id); //вызывается метод заполнения таблицы completed orders
 
-        return orderRepository.save(currentOrder);
+            return orderRepository.save(currentOrder);
+        }else{
+            log.warn("Заказ недоступен или уже выполняется");
+            throw new AccessDeniedException();
+        }
     }
